feat: add ability to track status of corpus for different analysis types
This commit enables tracking of the prepped status of a corpus per analysis type, such as countWords or countCollocations. The AnalysisType enum ensures the 4 analysis types are clear. EnumHelper.cpp ensures the AnalysisTypes can be converted into strings. JSONConvert makes sure that when the databaes returns no values that null can be returned for analysis_type and up_to_date which indicate to the front end that there is no data yet on the corpus status for this particular analysis type.